chore: move mocks to __mocks__, move imports
