+++ apps/web/components/CategoryTree/__tests__/CategoryTree.spec.ts
@@ -1,6 +1,6 @@
 import { mount } from '@vue/test-utils';
 import CategoryTree from '~/components/CategoryTree/CategoryTree.vue';
-import { CategoryMock } from '../../../__tests__/fixtures/category.mock';
+import { CategoryMock } from '../../../__tests__/__mocks__/category.mock';
 
 
 describe('<CategoryTree/>', () => {
