+++ apps/web/components/ProductAccordion/__tests__/ProductAccordion.spec.ts
@@ -1,5 +1,5 @@
 import { mount } from '@vue/test-utils';
-import { ProductMock } from '../../../__tests__/fixtures/product.mock';
+import { ProductMock } from '../../../__tests__/__mocks__/product.mock';
 import ProductAccordion from '~/components/ProductAccordion/ProductAccordion.vue';
