+++ apps/web/components/ui/PurchaseCard/__tests__/PurchaseCard.spec.ts
@@ -1,7 +1,7 @@
 import { mount } from '@vue/test-utils';
 import PurchaseCard from '~/components/ui/PurchaseCard/PurchaseCard.vue';
 import { ReviewAverage } from '@plentymarkets/shop-api';
-import { ProductMock } from '../../../../__tests__/fixtures/product.mock';
+import { ProductMock } from '../../../../__tests__/__mocks__/product.mock';
 
 describe('<PurchaseCard />', () => {
   it('should render component', () => {
